+++ argueview/visualizers/__init__.py
@@ -1,2 +1,3 @@
 from .initjs import *
 from .ToulminVisualizer import *
+from .FeatureListVisualizer import *